Move restore functionality to wizard initial state

+++ docker/trustpoint/wizard/transition/wizard_restore.sh
@@ -31,43 +31,64 @@ APACHE_TLS_DIR="/etc/trustpoint/tls"
 
 log INFO "STARTING APACHE RESTORE"
 
-# 5) (Re-)create and clean TLS directory
+# 5) Check wizard state
+log INFO "Checking for state file $WIZARD_INITIAL"
+if [ ! -f "$WIZARD_INITIAL" ]; then
+  log ERROR "State file $WIZARD_INITIAL not found"
+  exit 1
+fi
+
+# 6) (Re-)create and clean TLS directory
 log INFO "Ensuring TLS directory exists: $APACHE_TLS_DIR"
 mkdir -p "$APACHE_TLS_DIR"
 
 log INFO "Removing old TLS files from $APACHE_TLS_DIR"
 rm -f "$APACHE_TLS_DIR"/*
 
-# 6) Copy new TLS certs
+# 7) Copy new TLS certs
 log INFO "Copying new TLS files"
 cp /var/www/html/trustpoint/docker/trustpoint/apache/tls/* "$APACHE_TLS_DIR"/
 
-# 7) Remove existing Apache sites
+# 8) Remove existing Apache sites
 log INFO "Removing enabled Apache sites"
 rm -f /etc/apache2/sites-enabled/*
 
-# 8) Install HTTP config
+# 9) Install HTTP config
 log INFO "Installing HTTP config"
 cp /var/www/html/trustpoint/docker/trustpoint/apache/trustpoint-apache-http.conf \
    /etc/apache2/sites-available/
 cp /var/www/html/trustpoint/docker/trustpoint/apache/trustpoint-apache-http.conf \
    /etc/apache2/sites-enabled/
 
-# 9) Install HTTPS config
+# 10) Install HTTPS config
 log INFO "Installing HTTPS config"
 cp /var/www/html/trustpoint/docker/trustpoint/apache/trustpoint-apache-https.conf \
    /etc/apache2/sites-available/
 cp /var/www/html/trustpoint/docker/trustpoint/apache/trustpoint-apache-https.conf \
    /etc/apache2/sites-enabled/
 
-# 10) Enable Apache modules
+# 11) Enable Apache modules
 log INFO "Enabling Apache modules: ssl & rewrite"
 a2enmod ssl
 a2enmod rewrite
 
-# 11) Restart Apache gracefully
+# 12) Restart Apache gracefully
 log INFO "Restarting Apache (graceful)"
 apache2ctl graceful
+
+# Removes the current WIZARD_INITIAL state file.
+if ! rm "$WIZARD_INITIAL"
+then
+    log "ERROR: Failed to remove the WIZARD_INITIAL state file."
+    exit 3
+fi
+
+# Creates the WIZARD_COMPLETED state file.
+if ! touch "$WIZARD_COMPLETED"
+then
+    log "ERROR: Failed to create the WIZARD_COMPLETED state file."
+    exit 4
+fi
 
 log INFO "RESTORE SCRIPT COMPLETED SUCCESSFULLY"
 exit 0
